This is start of the merge bone system.
This is the start of the merge bones system, still needs a bit of work for it to be fully functional.

The basic comecpt that it will give of list of parent bones which are not in the exclude list. you then be able to merge a ratio of the bones which will help lower bones count for Pys bones in vrchat and etc.

Again this is basic and bare bones and needs work still.